% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CleanSeasonalData}
\alias{CleanSeasonalData}
\title{An importer function for NFL seasonal data.}
\usage{
CleanSeasonalData(year)
}
\arguments{
\item{a}{file name included in the data directory}
}
\description{
This function cleans one of the seasonal data files (included as a csv file in this package) and imports it into R.
}
\examples{
CleanSeasonalData()
}
\keyword{data}
\keyword{import;}

