% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CombineYearlyData}
\alias{CombineYearlyData}
\title{A function to combine multiple NFL seasonal data.}
\usage{
CombineYearlyData(years)
}
\arguments{
\item{a}{vector of years for NFL seasonal data to be imported into R.}
}
\description{
This function combines a vector of years corresponding to NFL seasonal data into a single R data frame.
}
\examples{
CombineYearlyData()
}
\keyword{combine;}
\keyword{data}
\keyword{seasonal}

